Converted the old compression code into the new ANT common compression class structure
git-svn-id: https://groke.otago.ac.nz/ant/ant@190 7489a7b4-4588-4892-8acd-774bacfab51c

+++ source/maths.h
@@ -27,7 +27,7 @@ inline int ANT_sign(double a) { return a < 0 ? -1 : a > 0 ? 1 : 0; }
 	ANT_FLOOR_LOG2()
 	----------------
 */
-inline unsigned long ANT_floor_log2(unsigned long x)
+inline unsigned long ANT_floor_log2(unsigned long long x)
 {
 extern unsigned long ANT_floor_log2_byte[];
 unsigned long sum, mult = 0;
